Cleanup AgPasture code. Remove bad met file.
