+++ Models/AgPasture/RootZone.cs
@@ -1,0 +1,12 @@
+﻿namespace Models.AgPasture
+{
+    using System;
+
+    [Serializable]
+    class RootZone
+    {
+        public string ZoneName;
+        public double RootDepth;
+        public double RootDM;
+    }
+}
