+++ Models/Models.csproj
@@ -132,9 +132,14 @@
     <Reference Include="System.Configuration" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="AgPasture\AgPasture.PastureSpecies.cs" />
-    <Compile Include="AgPasture\AgPasture.PastureSpecies.Organs.cs" />
     <Compile Include="AgPasture\AgPasture.Sward.cs" />
+    <Compile Include="AgPasture\PastureSpecies.cs" />
+    <Compile Include="AgPasture\PastureAboveGroundOrgan.cs" />
+    <Compile Include="AgPasture\PastureBelowGroundOrgan.cs" />
+    <Compile Include="AgPasture\GenericTissue.cs" />
+    <Compile Include="AgPasture\RootTissue.cs" />
+    <Compile Include="AgPasture\RootZone.cs" />
+    <Compile Include="AgPasture\SpeciesBasicStateSettings.cs" />
     <Compile Include="Agroforestry\AgroforestrySystem.cs" />
     <Compile Include="Agroforestry\LocalMicroClimate.cs" />
     <Compile Include="Agroforestry\TreeProxy.cs" />
